v1.1.0: rebrand, ParticleUtils
