+++ build.gradle
@@ -1,18 +1,6 @@
 import kr.entree.spigradle.attribute.*
 
-buildscript {
-    repositories {
-        maven {
-            url "https://plugins.gradle.org/m2/"
-        }
-    }
-    dependencies {
-        classpath "org.jlleitschuh.gradle:ktlint-gradle:9.2.1"
-    }
-}
-
 plugins {
-    id 'java'
     id 'org.jetbrains.kotlin.jvm' version '1.3.72'
     id 'kr.entree.spigradle' version '1.2.4'
 }
@@ -20,7 +8,10 @@ plugins {
 group 'com.dumbdogdiner'
 version '1.1.0'
 
+description 'Interactive Minecraft Parkour plugin with a personality.'
+
 sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 repositories {
     mavenCentral()
@@ -42,8 +33,8 @@ dependencies {
     compile 'com.okkero.skedule:skedule:1.2.6'
 
     implementation paper("1.15.2")
-    implementation vault()
 
+    implementation vault()
     implementation 'me.clip:placeholderapi:2.10.6'
 }
 
@@ -55,7 +46,7 @@ compileTestKotlin {
 }
 
 jar {
-    manifest { attributes 'Main-Class': 'Parkour.ParkourPluginKt' }
+    manifest { attributes 'Main-Class': 'Yapp.YappParkourPluginKt' }
     from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
 }
 
@@ -66,7 +57,7 @@ spigot {
     load = Load.STARTUP
     commands {
         create("parkour") {
-            aliases = List.of("pk")
+            aliases = List.of("pk", "yapp")
             description = "Manage parkour courses"
             usage = "/parkour [list/create/modify/delete]"
         }
