+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'Parkour'
+rootProject.name = 'yapp'
